17-02-22 add ft_putnbr_unsigned_fd to libft

+++ ft_printf_p.c
@@ -6,7 +6,7 @@
 /*   By: hbourgeo <hbourgeo@student.19.be>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/02/17 09:05:48 by hbourgeo          #+#    #+#             */
-/*   Updated: 2022/02/17 21:13:37 by hbourgeo         ###   ########.fr       */
+/*   Updated: 2022/02/18 08:09:14 by hbourgeo         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,10 +15,10 @@
 int ft_printf_p (va_list arg) 
 {
 	int	size;
-	unsigned long long int add;
+	unsigned long int add;
 
 	size = 0;
-	add = va_arg(arg, unsigned long long int);
+	add = va_arg(arg, unsigned long int);
 	if (add == 0)
 	{
 		size += ft_putstr_fd("0x0", 1);
